+++ classes/Controller.php
@@ -104,7 +104,7 @@ class Controller
         }
 
         if (!$this->redirect && isset($redirect)) {
-            $this->setRedirect($redirect);
+            $this->setRedirect($redirect, 303);
         }
 
         return $success;
@@ -140,25 +140,41 @@ class Controller
             return true;
         }
 
-        if ($this->authenticate($this->post)) {
+        // Remove login nonce from the form.
+        $form = array_diff_key($this->post, ['login-form-nonce' => true]);
+
+        // Fire Login process.
+        $event = $this->login->login($form, ['remember_me' => true], ['return_event' => true]);
+        $user = $event->getUser();
+
+        if ($user->authenticated) {
             $rateLimiter->resetRateLimit($ipKey, 'ip')->resetRateLimit($userKey);
 
-            $messages->add($t->translate('PLUGIN_LOGIN.LOGIN_SUCCESSFUL'), 'info');
+            $event->defMessage('PLUGIN_LOGIN.LOGIN_SUCCESSFUL', 'info');
 
-            $redirect = $this->grav['config']->get('plugins.login.redirect_after_login');
-            if (!$redirect) {
-                $redirect = $this->grav['session']->redirect_after_login ?: $this->grav['uri']->referrer('/');
-            }
-            $this->setRedirect($redirect);
+            $event->defRedirect(
+                $this->grav['session']->redirect_after_login
+                    ?: $this->grav['config']->get('plugins.login.redirect_after_login')
+                    ?: $this->grav['uri']->referrer('/')
+            );
         } else {
-            /** @var User $user */
-            $user = $this->grav['user'];
             if ($user->username) {
-                $messages->add($t->translate('PLUGIN_LOGIN.ACCESS_DENIED'), 'error');
-                $this->setRedirect($this->grav['config']->get('plugins.login.route_unauthorized', '/'));
+                $event->defMessage('PLUGIN_LOGIN.ACCESS_DENIED', 'error');
+
+                $event->defRedirect($this->grav['config']->get('plugins.login.route_unauthorized', '/'));
             } else {
-                $messages->add($t->translate('PLUGIN_LOGIN.LOGIN_FAILED'), 'error');
+                $event->defMessage('PLUGIN_LOGIN.LOGIN_FAILED', 'error');
             }
+        }
+
+        $message = $event->getMessage();
+        if ($message) {
+            $messages->add($t->translate($message), $event->getMessageType());
+        }
+
+        $redirect = $event->getRedirect();
+        if ($redirect) {
+            $this->setRedirect($redirect, $event->getRedirectCode());
         }
 
         return true;
@@ -171,9 +187,21 @@ class Controller
      */
     public function taskLogout()
     {
-        $this->login->logout(['remember_me' => true]);
+        $event = $this->login->logout(['remember_me' => true], ['return_event' => true]);
 
-        $this->setRedirect('/');
+        $message = $event->getMessage();
+        if ($message) {
+            /** @var Language $t */
+            $t = $this->grav['language'];
+
+            $messages = $this->grav['messages'];
+            $messages->add($t->translate($message), $event->getMessageType());
+        }
+
+        $redirect = $event->getRedirect() ?: $this->grav['config']->get('plugins.login.redirect_after_logout');
+        if ($redirect) {
+            $this->setRedirect($redirect, $event->getRedirectCode());
+        }
 
         return true;
     }
@@ -332,21 +360,6 @@ class Controller
     }
 
     /**
-     * Authenticate user.
-     *
-     * @param  array $form Form fields.
-     *
-     * @return bool
-     */
-    protected function authenticate($form)
-    {
-        // Remove login nonce.
-        $form = array_diff_key($form, ['login-form-nonce' => true]);
-
-        return $this->login->login($form, ['remember_me' => true])->authenticated;
-    }
-
-    /**
      * Redirects an action
      */
     public function redirect()
@@ -366,6 +379,14 @@ class Controller
     {
         $this->redirect = $path;
         $this->redirectCode = $code;
+    }
+
+    /**
+     * @return array Array containing [redirect, code].
+     */
+    public function getRedirect()
+    {
+        return [$this->redirect, $this->redirectCode];
     }
 
     /**
@@ -414,7 +435,7 @@ class Controller
      * @param  mixed $var A rememberMe instance to set
      *
      * @return RememberMe\RememberMe Returns the current rememberMe instance
-     * @deprecated 3.0 Use $grav['login']->rememberMe() instead
+     * @deprecated 2.5.0 Use $grav['login']->rememberMe() instead
      */
     public function rememberMe($var = null)
     {
@@ -431,7 +452,7 @@ class Controller
      * @param $count
      * @param $interval
      * @return bool
-     * @deprecated 3.0 Use $grav['login']->getRateLimiter($context) instead. See Grav\Plugin\Login\RateLimiter class.
+     * @deprecated 2.5.0 Use $grav['login']->getRateLimiter($context) instead. See Grav\Plugin\Login\RateLimiter class.
      */
     protected function isUserRateLimited(User $user, $field, $count, $interval)
     {
@@ -443,10 +464,27 @@ class Controller
      *
      * @param User $user
      * @param $field
-     * @deprecated 3.0 Use $grav['login']->getRateLimiter($context) instead. See Grav\Plugin\Login\RateLimiter class.
+     * @deprecated 2.5.0 Use $grav['login']->getRateLimiter($context) instead. See Grav\Plugin\Login\RateLimiter class.
      */
     protected function resetRateLimit(User $user, $field)
     {
         $this->login->resetRateLimit($user, $field);
+    }
+
+
+    /**
+     * Authenticate user.
+     *
+     * @param  array $form Form fields.
+     *
+     * @return bool
+     * @deprecated 2.6.2 Will be removed without replacement.
+     */
+    protected function authenticate($form)
+    {
+        // Remove login nonce.
+        $form = array_diff_key($form, ['login-form-nonce' => true]);
+
+        return $this->login->login($form, ['remember_me' => true])->authenticated;
     }
 }